+++ core/transfer/src/main/java/org/eclipse/dataspaceconnector/transfer/core/command/handlers/DeprovisionRequestHandler.java
@@ -39,4 +39,9 @@ public class DeprovisionRequestHandler extends SingleTransferProcessCommandHandl
         process.transitionDeprovisioning();
         return true;
     }
+
+    @Override
+    protected void postAction(TransferProcess process) {
+
+    }
 }