+++ core/transfer/src/main/java/org/eclipse/dataspaceconnector/transfer/core/command/handlers/SingleTransferProcessCommandHandler.java
@@ -44,6 +44,7 @@ public abstract class SingleTransferProcessCommandHandler<T extends SingleTransf
         } else {
             if (modify(transferProcess)) {
                 store.update(transferProcess);
+                postAction(transferProcess);
             }
         }
     }
@@ -57,4 +58,11 @@ public abstract class SingleTransferProcessCommandHandler<T extends SingleTransf
      * @return true if the process was actually modified, false otherwise.
      */
     protected abstract boolean modify(TransferProcess process);
+
+    /**
+     * Method that would be called after the entity update has been persisted.
+     *
+     * @param process the modified TransferProcess
+     */
+    protected abstract void postAction(TransferProcess process);
 }